Added tests for Sorghum BW5 and BW8
