+++ DCaPST.csproj
@@ -22,7 +22,7 @@
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
     <RegisterForComInterop>false</RegisterForComInterop>
-    <PlatformTarget>x86</PlatformTarget>
+    <PlatformTarget>AnyCPU</PlatformTarget>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DebugType>pdbonly</DebugType>
@@ -67,7 +67,9 @@
     <Compile Include="Canopy\LeafCanopy.cs" />
     <Compile Include="MethodExtensions\CCMMethodExtensions.cs" />
     <Compile Include="Output\LightVar.cs" />
-    <Compile Include="PathwayParameters\PathwayParamtersCCM.cs" />
+    <Compile Include="PathwayParameters\PathwayParametersC3.cs" />
+    <Compile Include="PathwayParameters\PathwayParametersC4.cs" />
+    <Compile Include="PathwayParameters\PathwayParametersCCM.cs" />
     <Compile Include="PhotosynthesisModel\PhotosynthesisModelCCM.cs" />
     <Compile Include="Tools\Lookup\LookupTable.cs" />
     <Compile Include="TemperatureFunctions\ModArrhehenius.cs" />
@@ -76,8 +78,6 @@
     <Compile Include="Layer\NaturalSpline.cs" />
     <Compile Include="Input\NotifyablePropertyClass.cs" />
     <Compile Include="PathwayParameters\PathwayParameters.cs" />
-    <Compile Include="PathwayParameters\PathwayParamtersC3.cs" />
-    <Compile Include="PathwayParameters\PathwayParamtersC4.cs" />
     <Compile Include="PhotosynthesisModel\PhotosynthesisModelC3.cs" />
     <Compile Include="PhotosynthesisModel\PhotosynthesisModel.cs" />
     <Compile Include="PhotosynthesisModel\PhotosynthesisModelC4.cs" />
